# app/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "customer API",
        "description": "This is the customer API",
        "version": "1.0.0"
    },
    "paths": {
        "/v1/customer": {
            "get": {
                "tags": [
                    "customer"
                ],
                "summary": "Retrieve customers information",
                "description": "list show paginated customers",
                "operationId": "customerIndex",
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Customer"
                                    }
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "User should be authorized to get profile information",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Not authorized"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearer": []
                    }
                ]
            }
        },
        "/v1/customer/": {
            "post": {
                "tags": [
                    "customer"
                ],
                "summary": "create customer",
                "description": "create customer by given information",
                "operationId": "customerCreate",
                "requestBody": {
                    "description": "Pass user credentials",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "email"
                                ],
                                "properties": {
                                    "first_name": {
                                        "type": "string",
                                        "example": "user1"
                                    },
                                    "last_name": {
                                        "type": "string",
                                        "example": "user1"
                                    },
                                    "email": {
                                        "type": "string",
                                        "format": "email",
                                        "example": "user1@mail.com"
                                    },
                                    "phone_number": {
                                        "type": "string",
                                        "example": "+989189151266"
                                    },
                                    "bank_account_number": {
                                        "type": "string",
                                        "example": "51655555000555"
                                    },
                                    "date_of_birth": {
                                        "type": "string",
                                        "format": "date",
                                        "example": "1999-11-05"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "$ref": "#/components/schemas/Customer"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "User should be authorized to get profile information",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Not authorized"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearer": []
                    }
                ]
            }
        },
        "/v1/customer/{id}": {
            "get": {
                "tags": [
                    "customer"
                ],
                "summary": "Retrieve customer information by id",
                "description": "show customer information",
                "operationId": "customerShow",
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "$ref": "#/components/schemas/Customer"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "User should be authorized to get profile information",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Not authorized"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearer": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "customer"
                ],
                "summary": "update customer information",
                "description": "update customer information by id",
                "operationId": "customerUpdate",
                "requestBody": {
                    "description": "Pass user credentials",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "email"
                                ],
                                "properties": {
                                    "first_name": {
                                        "type": "string",
                                        "example": "user1"
                                    },
                                    "last_name": {
                                        "type": "string",
                                        "example": "user1"
                                    },
                                    "email": {
                                        "type": "string",
                                        "format": "email",
                                        "example": "user1@mail.com"
                                    },
                                    "phone_number": {
                                        "type": "string",
                                        "example": "+989189151266"
                                    },
                                    "bank_account_number": {
                                        "type": "string",
                                        "example": "51655555000555"
                                    },
                                    "date_of_birth": {
                                        "type": "string",
                                        "format": "date",
                                        "example": "1999-11-05"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "$ref": "#/components/schemas/Customer"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "User should be authorized to get profile information",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Not authorized"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearer": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "customer"
                ],
                "summary": "remove customer information",
                "description": "remove customer information by id",
                "operationId": "customerDelete",
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "OK"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "User should be authorized to get profile information",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Not authorized"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearer": []
                    }
                ]
            }
        }
    },
    "components": {
        "schemas": {
            "Customer": {
                "required": [
                    "email"
                ],
                "properties": {
                    "id": {
                        "type": "integer",
                        "readOnly": "true",
                        "example": "1"
                    },
                    "first_name": {
                        "type": "string",
                        "maxLength": 50,
                        "example": "John"
                    },
                    "last_name": {
                        "type": "string",
                        "maxLength": 50,
                        "example": "Doe"
                    },
                    "email": {
                        "description": "Customer unique email address",
                        "type": "string",
                        "format": "email",
                        "example": "customer@gmail.com"
                    },
                    "phone_number": {
                        "description": "Customer phone number",
                        "type": "string"
                    },
                    "bank_account_number": {
                        "description": "bank account number",
                        "type": "string",
                        "format": "string"
                    },
                    "date_of_birth": {
                        "description": "Date of birth customer",
                        "type": "string",
                        "format": "date",
                        "example": "1990-02-25"
                    },
                    "created_at": {
                        "description": "Initial creation timestamp",
                        "type": "string",
                        "format": "date-time",
                        "readOnly": "true"
                    },
                    "updated_at": {
                        "description": "Last update timestamp",
                        "type": "string",
                        "format": "date-time",
                        "readOnly": "true"
                    }
                },
                "type": "object",
                "xml": {
                    "name": "Customer"
                }
            }
        }
    }
}